+++ NEWS.md
@@ -1,3 +1,10 @@
+# climate 1.2.0
+
+* Support for telemetry dataset from IMGW-PIB repository
+    * `data.table` package used for efficient processing of large amount of data in this subrepository
+* Update for location of IMGW-PIB telemetry stations
+
+
 # climate 1.1.1
 
 * Fix problems with downloading `precip` dataset from IMGW-PIB repository after recent changes in metadata